/libs directory code now has sufficient comments.
Each library that implements a generic data structure now has a
high-level description, asymptotic run-time descriptions, and usage at
the top of the header, and each method implemented commented with the
meaning of the arguments it takes and the values it returns. This whole
generic-container-using-macros thing has just passed proof-of-concept
stage.