Added capability for error reporting by line number
lexer.c and lexer.h modified to reflect changes. errormodule.h is new,
defines fileLoc, which is a "universal file position locator" of sorts,
containing a FILE*, the string representing the file path and name, a
line number and a column number within the file. get_char() and
unget_char() are wrappers over getc and ungetc, respectively, to
respect position information.